requirements.txt
beautifulsoup4
tabula-py
python-dotenv
requests
tweepy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
